requirements.txt
pyyaml>=6.0
requests>=2.28
dnspython>=2.3
rich>=13.0
paramiko>=3.0
